Reading directly from lambda.tf in the generate appspec script
