+++ buildspec.yml
@@ -13,7 +13,7 @@ phases:
   post_build:
     commands:
       - echo Build completed on `date`
-      - pip install boto3 pyyaml
+      - pip install boto3 pyyaml python-hcl2
       - python3 generate_appspec.py
 
 artifacts:
