+++ test/requirements.txt
@@ -1,3 +1,4 @@
-python-hcl2>=0.3.0
+boto3>=1.26.0
 pyyaml>=6.0
-boto3>=1.26.0 
+python-hcl2>=0.3.0
+pytest>=7.0.0
